requirements-docker.txt
# وابستگی‌های اجرای API در Docker — بدون dev/e2e (playwright، pytest)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
redis>=5.0.1
celery>=5.3.6
jinja2>=3.1.3
jdatetime>=4.1.0
openpyxl>=3.1.0
fpdf2>=2.7.0
arabic-reshaper>=3.0.0
python-bidi>=0.4.2
